Move isNip05Verfied to NostrUtil
